fixes #32. Handle misformed year in pdf name (just appendings/prependings)
code will compare year in pdf name and current year and will take the
current year if something is appended or prepended to the year in the
pdf. This is a quite hacky solution and does not cover typing errors
like when someone types 2017 when its 2018 or so.
Probably a better idea is just take current year except when its end of
december.
But it fixes the problem of week 14/15 in 2018 when the pdf name is:

Garching_Speiseplan_KW_14_20181
Garching_Speiseplan_KW_15_20181