IPP correct coulumn detection with shift -3; fixes #28
The column detection is done through the string "Tagessuppe siehe Aushang" which is at the beginning of
every column. However, due to center alignment the column do not begin at the 'T' character and broader
text in the column might be left of this character, which then gets truncated. But the gap between the 'T'
and the '€' character of the previous column — the real beginning of the current column — is always three,
is will be subtracted here. Monday is the second column, so the value should never become negative although
it is handled here.