IPP correct biginning of column, shift -3 (#37)
* FMI pdf change KW03 → KW_03, but use regex for much more cases

* regex for pdf of IPP

* transform the supported canteen list into a more clearer table

* add telegram channel to projects using eat-api

* hard code google maps positon of ipp since search is too unprecise

* partly fixes #32. FMI loop for multiple pdf on homepage

When multiple PDF on the homepage are published, a loop now goes
through all pdfs and will extract the menus.

* fixes #32. Handle misformed year in pdf name (just appendings/prependings)

code will compare year in pdf name and current year and will take the
current year if something is appended or prepended to the year in the
pdf. This is a quite hacky solution and does not cover typing errors
like when someone types 2017 when its 2018 or so.
Probably a better idea is just take current year except when its end of
december.
But it fixes the problem of week 14/15 in 2018 when the pdf name is:

Garching_Speiseplan_KW_14_20181
Garching_Speiseplan_KW_15_20181

* check for Speiseplan in href url instead of text

The text of the a tag can contain an node again (such as the strong tag)
the xpath contain, then fails since it is a new sub node.

* Closed days fix ipp (#34)

* FMI pdf change KW03 → KW_03, but use regex for much more cases

* regex for pdf of IPP

* transform the supported canteen list into a more clearer table

* add telegram channel to projects using eat-api

* hard code google maps positon of ipp since search is too unprecise

* partly fixes #32. FMI loop for multiple pdf on homepage

When multiple PDF on the homepage are published, a loop now goes
through all pdfs and will extract the menus.

* fixes #32. Handle misformed year in pdf name (just appendings/prependings)

code will compare year in pdf name and current year and will take the
current year if something is appended or prepended to the year in the
pdf. This is a quite hacky solution and does not cover typing errors
like when someone types 2017 when its 2018 or so.
Probably a better idea is just take current year except when its end of
december.
But it fixes the problem of week 14/15 in 2018 when the pdf name is:

Garching_Speiseplan_KW_14_20181
Garching_Speiseplan_KW_15_20181

* check for Speiseplan in href url instead of text

The text of the a tag can contain an node again (such as the strong tag)
the xpath contain, then fails since it is a new sub node.

* Allows to parse pdf with empty column (closed days)

* remove debugging print

* IPP test cases for weeks with holidays

* Revert "Closed days fix ipp (#34)"

This reverts commit 2a02f9a11c8677d54a70376ec1693032f7346b04.

* Closed days fix ipp (reopen) (#35)

* FMI pdf change KW03 → KW_03, but use regex for much more cases

* regex for pdf of IPP

* transform the supported canteen list into a more clearer table

* add telegram channel to projects using eat-api

* hard code google maps positon of ipp since search is too unprecise

* partly fixes #32. FMI loop for multiple pdf on homepage

When multiple PDF on the homepage are published, a loop now goes
through all pdfs and will extract the menus.

* fixes #32. Handle misformed year in pdf name (just appendings/prependings)

code will compare year in pdf name and current year and will take the
current year if something is appended or prepended to the year in the
pdf. This is a quite hacky solution and does not cover typing errors
like when someone types 2017 when its 2018 or so.
Probably a better idea is just take current year except when its end of
december.
But it fixes the problem of week 14/15 in 2018 when the pdf name is:

Garching_Speiseplan_KW_14_20181
Garching_Speiseplan_KW_15_20181

* check for Speiseplan in href url instead of text

The text of the a tag can contain an node again (such as the strong tag)
the xpath contain, then fails since it is a new sub node.

* Allows to parse pdf with empty column (closed days)

* remove debugging print

* IPP test cases for weeks with holidays

* IGNORECASE for IPP regex

* warning if not all 5 colums are detected. Turn it into exception?

* IPP correct coulumn detection with shift -3; fixes #28

The column detection is done through the string "Tagessuppe siehe Aushang" which is at the beginning of
every column. However, due to center alignment the column do not begin at the 'T' character and broader
text in the column might be left of this character, which then gets truncated. But the gap between the 'T'
and the '€' character of the previous column — the real beginning of the current column — is always three,
is will be subtracted here. Monday is the second column, so the value should never become negative although
it is handled here.

* comment typo