IPP search for the soup line instead assuming fixed position (#38)
* FMI pdf change KW03 → KW_03, but use regex for much more cases

* regex for pdf of IPP

* transform the supported canteen list into a more clearer table

* add telegram channel to projects using eat-api

* hard code google maps positon of ipp since search is too unprecise

* partly fixes #32. FMI loop for multiple pdf on homepage

When multiple PDF on the homepage are published, a loop now goes
through all pdfs and will extract the menus.

* fixes #32. Handle misformed year in pdf name (just appendings/prependings)

code will compare year in pdf name and current year and will take the
current year if something is appended or prepended to the year in the
pdf. This is a quite hacky solution and does not cover typing errors
like when someone types 2017 when its 2018 or so.
Probably a better idea is just take current year except when its end of
december.
But it fixes the problem of week 14/15 in 2018 when the pdf name is:

Garching_Speiseplan_KW_14_20181
Garching_Speiseplan_KW_15_20181

* check for Speiseplan in href url instead of text

The text of the a tag can contain an node again (such as the strong tag)
the xpath contain, then fails since it is a new sub node.

* Closed days fix ipp (#34)

* FMI pdf change KW03 → KW_03, but use regex for much more cases

* regex for pdf of IPP

* transform the supported canteen list into a more clearer table

* add telegram channel to projects using eat-api

* hard code google maps positon of ipp since search is too unprecise

* partly fixes #32. FMI loop for multiple pdf on homepage

When multiple PDF on the homepage are published, a loop now goes
through all pdfs and will extract the menus.

* fixes #32. Handle misformed year in pdf name (just appendings/prependings)

code will compare year in pdf name and current year and will take the
current year if something is appended or prepended to the year in the
pdf. This is a quite hacky solution and does not cover typing errors
like when someone types 2017 when its 2018 or so.
Probably a better idea is just take current year except when its end of
december.
But it fixes the problem of week 14/15 in 2018 when the pdf name is:

Garching_Speiseplan_KW_14_20181
Garching_Speiseplan_KW_15_20181

* check for Speiseplan in href url instead of text

The text of the a tag can contain an node again (such as the strong tag)
the xpath contain, then fails since it is a new sub node.

* Allows to parse pdf with empty column (closed days)

* remove debugging print

* IPP test cases for weeks with holidays

* Revert "Closed days fix ipp (#34)"

This reverts commit 2a02f9a11c8677d54a70376ec1693032f7346b04.

* Closed days fix ipp (reopen) (#35)

* FMI pdf change KW03 → KW_03, but use regex for much more cases

* regex for pdf of IPP

* transform the supported canteen list into a more clearer table

* add telegram channel to projects using eat-api

* hard code google maps positon of ipp since search is too unprecise

* partly fixes #32. FMI loop for multiple pdf on homepage

When multiple PDF on the homepage are published, a loop now goes
through all pdfs and will extract the menus.

* fixes #32. Handle misformed year in pdf name (just appendings/prependings)

code will compare year in pdf name and current year and will take the
current year if something is appended or prepended to the year in the
pdf. This is a quite hacky solution and does not cover typing errors
like when someone types 2017 when its 2018 or so.
Probably a better idea is just take current year except when its end of
december.
But it fixes the problem of week 14/15 in 2018 when the pdf name is:

Garching_Speiseplan_KW_14_20181
Garching_Speiseplan_KW_15_20181

* check for Speiseplan in href url instead of text

The text of the a tag can contain an node again (such as the strong tag)
the xpath contain, then fails since it is a new sub node.

* Allows to parse pdf with empty column (closed days)

* remove debugging print

* IPP test cases for weeks with holidays

* IGNORECASE for IPP regex

* warning if not all 5 colums are detected. Turn it into exception?

* IPP search for the soup line instead assuming fixed position

A sligthly changed pdf format will add some new empty lines and breaks
the parsing.

`lines` when used to construct `postitions1` and `positions2`
looks now like:

    ['', '', '                        Tagessuppe siehe Aushang                   Tagessuppe siehe Aushang   …

e.g. the two new empty lines at the biginning are new.
To be risilient against further changes the "Tagessuppe" line will be
searchd now intead of assuming a fixed line position.